# ARROW-9919: [Rust][DataFusion] Speedup math operations by 15%+