This PR improves speed of all math operations by 15% (on default batch size), and changes how these operations scale with batch size (to be faster).

See main issue here: ARROW-9918

This has a big contribution from @yordan-pavlov, that initially reported the issue on ARROW-8908 for literal arrays.

Benchmarks (on my computer, before and after the second commit):

```
sqrt_20_12              time:   [34.422 ms 34.503 ms 34.584 ms]                       
                        change: [-16.333% -16.055% -15.806%] (p = 0.00 < 0.05)
                        Performance has improved.

sqrt_22_12              time:   [150.13 ms 150.79 ms 151.42 ms]                       
                        change: [-16.281% -15.488% -14.779%] (p = 0.00 < 0.05)
                        Performance has improved.

sqrt_22_14              time:   [151.45 ms 151.68 ms 151.90 ms]                       
                        change: [-18.233% -16.919% -15.888%] (p = 0.00 < 0.05)
                        Performance has improved.
```