@paddyhoran 2020-09-07
Hey @jorgecarleitao,

Sorry, my info was wrong.  As you said `<From<Vec<_>>` does do a full copy of the underlying data which uses memory.rs [here](https://github.com/apache/arrow/blob/master/rust/arrow/src/buffer.rs#L274).  I was not aware of this.  

I think you are right that there are a lot of improvements to be made to the builders, we should be able to get better performance out of them as they should be able to act like a `Vec` where the correct alignment is enforced so that no copy is needed.

Anyway, sorry for the noise and thank you for the detailed explanation.