@nevi-me 2020-09-07
@paddyhoran @jorgecarleitao @andygrove  PTAL at my suggestion before merging (https://github.com/apache/arrow/pull/8116#discussion_r484493616)